send_serial_port/requirements.txt
pyserial~=3.5
pyinstaller